vcs/gitcmd: ResolveRevision: Use "^{commit}" peeling operator to detect missing objects.
This will ensure that the output actually names an object in the object
database and/or can be used as the specific type of object we require
(i.e., a commit, branch, tag, or an annotated tag that points at a
commit).
See details at http://git-scm.com/docs/git-rev-parse.