Add BranchesOptions, support getting behind/ahead counts.
This is a breaking API change for Branches method of Repository
interface.
It changes the signature of Branches to be more like Commits, which
accepts an Options struct where parameters may be specified that
describe details of the Branches query.
It also adds a total uint return value, which will always be the total
number of branches, even if fewer are returned due to pagination, for
example. Note that there is no pagination support at this time, but
since it's likely to be added in the future, it's hopefully better to
avoid a second API breakage at that time.
Implement support for BehindAheadBranch in vcs/gitcmd implementation.
If set to a non-empty string, it uses that branch name as the base
branch, and gets behind/ahead commit counts for each branch returned.