README: Improve installation steps description.
Mention that both `libssh2` and `libgcrypt` are hard prerequisites for
libgit2 implementation in advance of its installation.
The Makefile "install" target was outdated (missing the `git checkout
next` step). In addition to that, it repeats the exact same steps as
described in Resolving Dependencies section, which is assumed to be
done by the time user gets to Installing section. I removed it since
it's unneeded and may cause confusion.
Closes #43.