vcs/*: fix handling of root paths inside VFS backends
This change handles root paths (i.e. "/foo") as if they were relative
("foo") such that asking the VFS for a file at the root works properly.

Without this change it is impossible to use the VFS provided with:

  golang.org/x/tools/godoc/vfs/httpfs

As net/http.FileSystem will implicitly convert all URL paths into
absolute paths at the root of the FS.

Additionally this makes the behavior of the VFS backends comply with
the OS VFS implementation provided by the godoc/vfs package.

Fixes #23