Export environ.Unset method to make its API nicer.
I originally made it unexported since it didn't need to be exported
(it's a private internal helper for its package only). However, it's
already unexported by virtual of being a method of an unexported type.

Making it exported makes more sense since it's a part of the method's
API, meant to be used by external callers. It's better to keep
unexported methods for internal methods (not part of the type's API).