Change vcs.ListFiles interface to also guarantee sort order.
The current only implementation already returns results sorted
alphabetically. Upgrade the interface to make this a part of the
contract.

If in the future other implementations have reasons to return unsorted
results, we can downgrade this guarantee appropriately, but until then,
we might as well match reality to avoid users having to sort output a
second time.

@pararthshah, please review and see if you agree with making this
interface contract change.