vcs/util/tracer: add a utility package for recording VCS operations to Appdash
This change adds a utility package which records VCS operations to an `*appdash.Recorder`
by wrapping a `vcs.Repository` value. It implements the optional interface types
that a `vcs.Repository` implementation may opt-in to, so no functionality is
lost. Because it works at the interface level we can trace all different
backends (`gitcmd`, `git`, `hg`, etc) without writing new tracing code.

In Appdash, all events are prefixed with `vcs.Type` for easier filtering:

![screenshot](http://virtivia.com:27080/1nt70brhgfma7.png)