vcs/gitcmd: Use correct commit range in commitLog.
vcs.CommitsOptions has Head and Base fields, documented as:

	Head CommitID // include all commits reachable from this commit (required)
	Base CommitID // exclude all commits reachable from this commit (optional, like `git log Base..Head`)

The Base field is documented to produce results equivalent to
git log Base..Head. However, that was not the case, the previous
code produced results equivalent to git log Head...Base instead.

In many cases, those two are equivalent, which is likely why it
has gone unnoticed (and the previous tests didn't catch this).

Add more thorough test coverage to catch the subtle difference.