vcs/git, vcs/gitcmd: Always dereference annotated tags.
This change fixes the inconsistency in annotated tag dereferencing
between various git backends by opting to always dereference.

This behavior is more consistent with the high level nature of the
vcs API, and it's easier for users to consume. If detailed tag
information is needed, support for that can be added as an optional
enhancement.

Document showRef so its behavior is specified and understood.
Don't sort via byteSlices, because git show-ref output is
already sorted. Add --deference flag, and make use of it
in Repository.Tags method. It has no effect for non-tags.

Enable annotated tag test cases in TestRepository_FileSystem_fromTag,
they are passing now.

Fixes #24.