vcs/gitcmd: not use --follow for glob pathspecs
This change is a followup to PR #50, which added support for
constraining history to a file/dir path.

In that PR, --follow flag is provided git log invocation in
Repository.commitLog method whenever opt.Path is non-empty.

However, the current version of git has some constraints on when
the --follow flag¹ may be used:

	--follow
	  Continue listing the history of a file beyond renames (works
	  only for a single file).

Unfortunately, parsing a git pathspec² is not easily implemented,
so the current solution is to detect one specific instance of a
pathspec that does not permit use of --follow flag, and not use
the --follow flag in that case. When there's a need for handling
other pathspec magic types correctly, this fix can be adjusted
further accordingly.

Also modify the text of an error in commitLog to provide more
information about the command that failed to run, and be more
consistent with other similar errors in this .go file.

¹ https://git-scm.com/docs/git-log/2.20.1#git-log---follow
  https://github.com/git/git/blob/v2.20.1/tree-diff.c#L585-L590
² https://git-scm.com/docs/gitglossary/2.20.1#def_pathspec
  https://github.com/git/git/blob/v2.20.1/tree-diff.c#L585-L590

Fixes #110